Website/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.32.3
sqlalchemy==2.0.30
rapidfuzz==3.6.2
redis==5.0.4
numpy==1.26.4
chromadb==0.5.0
sentence-transformers==3.0.1
tqdm==4.66.4
